requirements.txt
opencv-python==4.1.0.25
numpy==1.17.0
Pillow==6.1.0
click==6.7